refactor: isolate lazy route entry handling
