+++ src/lazy-compilation.ts
@@ -1,5 +1,9 @@
-import { BUILD_CLIENT_ROUTE_QUERY_STRING } from './constants.js';
-import type { PluginOptions } from './types.js';
+import type { Rspack } from '@rsbuild/core';
+import {
+  BUILD_CLIENT_ROUTE_QUERY_STRING,
+  CLIENT_ROUTE_EXPORTS_SET,
+} from './constants.js';
+import type { PluginOptions, Route } from './types.js';
 
 type LazyCompilationOptions = Exclude<
   NonNullable<PluginOptions['lazyCompilation']>,
@@ -19,6 +23,9 @@ const normalizeSlashes = (value: string): string => value.replace(/\\/g, '/');
 const STYLE_RESOURCE_RE =
   /\.(?:css|less|sass|scss|styl|stylus|pcss|postcss|sss)(?:[?#!]|$)/i;
 const NATIVE_STYLE_RESOURCE_PATTERN = String.raw`\.(?:css|less|sass|scss|styl|stylus|pcss|postcss|sss)(?:[?#!].*)?$`;
+const LAZY_ROUTE_ENTRY_EXECUTION =
+  /var __webpack_exports__ = __webpack_exec__\([^)]*!lazy-compilation-proxy/;
+const LAZY_ROUTE_EXPORTS_IDENTIFIER = '__reactRouterLazyRouteExports';
 
 const escapeRegExp = (value: string): string =>
   value.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
@@ -171,4 +178,103 @@ export const guardReactRouterLazyCompilation = ({
       return matchesLazyCompilationTest(userTest, lazyModule);
     },
   };
+};
+
+export const isLazyRouteEntrySource = (source: string): boolean =>
+  LAZY_ROUTE_ENTRY_EXECUTION.test(source);
+
+export const createLazyRouteEntryExportBridge = (
+  exportNames: readonly string[]
+): string => {
+  const clientExportNames = Array.from(
+    new Set(
+      exportNames.filter(exportName => CLIENT_ROUTE_EXPORTS_SET.has(exportName))
+    )
+  );
+  if (clientExportNames.length === 0) {
+    return '';
+  }
+
+  const bindings = clientExportNames.map(
+    (exportName, index) =>
+      `const __reactRouterLazyRouteExport_${index} = ${LAZY_ROUTE_EXPORTS_IDENTIFIER}[${JSON.stringify(
+        exportName
+      )}];`
+  );
+  const exports = clientExportNames.map(
+    (exportName, index) =>
+      `__reactRouterLazyRouteExport_${index} as ${exportName}`
+  );
+
+  return [
+    `const ${LAZY_ROUTE_EXPORTS_IDENTIFIER} = await __webpack_exports__;`,
+    ...bindings,
+    `export { ${exports.join(', ')} };`,
+  ].join('\n');
+};
+
+export const appendLazyRouteEntryExportBridges = ({
+  assets,
+  compilation,
+  moduleExportsByRouteId,
+  routes,
+  sources,
+  stats,
+}: {
+  assets: Record<string, Rspack.sources.Source>;
+  compilation: Rspack.Compilation;
+  moduleExportsByRouteId: Readonly<Record<string, readonly string[]>>;
+  routes: Record<string, Route>;
+  sources: Pick<typeof Rspack.sources, 'ConcatSource'>;
+  stats: { assetsByChunkName?: Record<string, string[]> } | undefined;
+}): void => {
+  for (const [routeId, route] of Object.entries(routes)) {
+    const extensionIndex = route.file.lastIndexOf('.');
+    if (extensionIndex === -1) {
+      continue;
+    }
+    const entryName = route.file.slice(0, extensionIndex);
+    const routeEntryAssets = stats?.assetsByChunkName?.[entryName];
+    const routeEntryAssetName = routeEntryAssets?.find(assetName =>
+      assetName.endsWith(`${entryName}.js`)
+    );
+    if (!routeEntryAssetName) {
+      if (routeEntryAssets && routeEntryAssets.length > 0) {
+        throw new Error(
+          `[rsbuild-plugin-react-router] Expected a JavaScript route-entry asset for lazy route \`${routeId}\`.`
+        );
+      }
+      continue;
+    }
+
+    const source = assets[routeEntryAssetName];
+    if (!source) {
+      throw new Error(
+        `[rsbuild-plugin-react-router] Could not read lazy route-entry asset \`${routeEntryAssetName}\`.`
+      );
+    }
+    const sourceText = source.source().toString();
+    if (!sourceText) {
+      throw new Error(
+        `[rsbuild-plugin-react-router] Lazy route-entry asset \`${routeEntryAssetName}\` was empty.`
+      );
+    }
+    if (
+      !isLazyRouteEntrySource(sourceText) ||
+      sourceText.includes(LAZY_ROUTE_EXPORTS_IDENTIFIER)
+    ) {
+      continue;
+    }
+
+    const exportBridge = createLazyRouteEntryExportBridge(
+      moduleExportsByRouteId[routeId] ?? []
+    );
+    if (!exportBridge) {
+      continue;
+    }
+    compilation.updateAsset(
+      routeEntryAssetName,
+      new sources.ConcatSource(source, '\n', exportBridge, '\n')
+    );
+  }
 };
